refactor(svg-icons): packaging
* Produce UMD and ES bundles using Rollup
* Moved source files to `src`
* Added JSDoc annotations
* Marked package as side-effects free
