--- packages/svg-icons/icons/svgIcon.interface.ts
@@ -1,5 +1,0 @@
-export interface SVGIcon {
-    name: string;
-    content: string;
-    viewBox: string;
-}
