--- packages/svg-icons/index.ts
@@ -1,6 +1,0 @@
-export { arrowUp } from './icons/arrowUp';
-export { arrowRight } from './icons/arrowRight';
-export { arrowDown } from './icons/arrowDown';
-export { arrowLeft } from './icons/arrowLeft';
-
-export { caretTr } from './icons/caretTr';
