+++ packages/svg-icons/src/icons/arrow-right.ts
@@ -1,4 +1,4 @@
-import { SVGIcon } from './svgIcon.interface';
+import { SVGIcon } from '../svg-icon.interface';
 
 export const arrowRight: SVGIcon = {
     name: 'arrow-right',
