+++ packages/svg-icons/src/index.ts
@@ -1,0 +1,8 @@
+export { SVGIcon } from './svg-icon.interface';
+
+export { arrowUp } from './icons/arrow-up';
+export { arrowRight } from './icons/arrow-right';
+export { arrowDown } from './icons/arrow-down';
+export { arrowLeft } from './icons/arrow-left';
+
+export { caretTr } from './icons/caret-tr';
