+++ packages/svg-icons/tsconfig.es.json
@@ -1,26 +1,6 @@
 {
+    "extends": "./tsconfig.json",
     "compilerOptions": {
-        "outDir": "dist/es",
-        "target": "es5",
-        "module": "es2015",
-        "moduleResolution": "node",
-        "lib": ["es2015", "dom"],
-        "types": [],
-
-        "declaration": false,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "importHelpers": true,
-        "noImplicitAny": false,
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "removeComments": false
-    },
-    "files": [
-        "index.ts"
-    ],
-    "include": [],
-    "exclude": [
-        "node_modules"
-    ]
+        "target": "es2015"
+    }
 }
