+++ packages/svg-icons/tsconfig.json
@@ -1,25 +1,19 @@
 {
     "compilerOptions": {
-        "outDir": "dist/",
+        "outDir": "dist",
         "target": "es5",
-        "module": "commonjs",
+        "module": "es2015",
         "lib": ["es2015", "dom"],
         "types": [],
 
-        "declaration": false,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "importHelpers": true,
+        "declaration": true,
+        "importHelpers": false,
         "noImplicitAny": false,
         "noUnusedLocals": true,
         "noUnusedParameters": true,
         "removeComments": false
     },
-    "files": [
-        "index.ts"
-    ],
-    "include": [],
-    "exclude": [
-        "node_modules"
+    "include": [
+        "src/**/*.ts"
     ]
 }
